report/requirements.txt
pandas
redshift_connector
python-dotenv